main/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
h5py>=3.7.0
opencv-python>=4.6.0
Pillow>=9.0.0
pyyaml>=6.0
tqdm>=4.64.0
scikit-learn>=1.1.0
